Move processor from annotations to processor package
