+++ src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-com.j256.ormlite.android.annotations.OrmLiteAnnotationProcessor
+com.j256.ormlite.android.processor.OrmLiteAnnotationProcessor
